[Update] Remove apk update from Dockerfile  (#418)
* Update Dockerfile.

Remove `--update` to not use the latest apk index to accelerate the docker pull and build, and also make the version more reliable. Tried to work around a potential problem where the ruby dependency is incorrect while we can sill build mdl, but that didn't work out. Things I tried are listed below…

- Simply use Ruby without dependencies.
- Try adding bundler to see if it help add mdl.
- Try using rake for building mdl.
- Try including libs to allow building.
- Finally retreat to using devs for building the gem. Ah!

Back to basics. Author fixed the dependency.